+++ include/flecs/addons/cpp/utils/enum.hpp
@@ -96,7 +96,7 @@ struct enum_type {
     }
 
     flecs::entity_t entity(E value) const {
-        return data.constants[value].id;
+        return data.constants[static_cast<int>(value)].id;
     }
 
 private:
